feat(Deployments): delete LoC and add typescript support
