--- constants.js
@@ -1,13 +1,0 @@
-const DRAW_BUFFER_CARDINALITY = 255; 
-const PRIZE_DISTRIBUTION_BUFFER_CARDINALITY = 8; 
-const BEACON_PERIOD_SECONDS = 60 * 20; 
-const DRAW_CALCULATOR_TIMELOCK = 60 * 10; 
-const TOKEN_DECIMALS = 6;
-
-module.exports = {
-  DRAW_BUFFER_CARDINALITY,
-  PRIZE_DISTRIBUTION_BUFFER_CARDINALITY,
-  BEACON_PERIOD_SECONDS,
-  DRAW_CALCULATOR_TIMELOCK,
-  TOKEN_DECIMALS
-}
