+++ constants.ts
@@ -1,0 +1,13 @@
+export const DRAW_BUFFER_CARDINALITY: Number = 255;
+export const PRIZE_DISTRIBUTION_BUFFER_CARDINALITY: Number = 8;
+export const BEACON_PERIOD_SECONDS: Number = 60 * 20;
+export const DRAW_CALCULATOR_TIMELOCK: Number = 60 * 10;
+export const TOKEN_DECIMALS: Number = 6;
+
+export default {
+  DRAW_BUFFER_CARDINALITY,
+  PRIZE_DISTRIBUTION_BUFFER_CARDINALITY,
+  BEACON_PERIOD_SECONDS,
+  DRAW_CALCULATOR_TIMELOCK,
+  TOKEN_DECIMALS
+}
