+++ deploy/mumbai.ts
@@ -1,15 +1,14 @@
-const { dim, cyan, green } = require('../colors')
-const { deployContract } = require('../deployContract')
 
-const { 
+import { dim, cyan, green } from '../utils/colors'
+import { deployContract } from '../utils/deployContract'
+import {
   DRAW_BUFFER_CARDINALITY,
   PRIZE_DISTRIBUTION_BUFFER_CARDINALITY,
   DRAW_CALCULATOR_TIMELOCK,
-  TOKEN_DECIMALS 
-} = require('../constants')
+  TOKEN_DECIMALS
+} from '../constants'
 
-module.exports = async (hardhat) => {
-
+export default async (hardhat: any) => {
   const {
     ethers,
     deployments,
@@ -21,7 +20,7 @@ module.exports = async (hardhat) => {
     deployer,
     manager
   } = await getNamedAccounts();
-  
+
   if (process.env.DEPLOY != 'mumbai') {
     dim(`Ignoring mumbai...`)
     return
@@ -30,10 +29,18 @@ module.exports = async (hardhat) => {
   }
 
   const mockYieldSourceResult = await deployContract(deploy, 'MockYieldSource', deployer, ['Token', 'TOK', TOKEN_DECIMALS])
-  const yieldSourcePrizePoolResult = await deployContract(deploy, 'YieldSourcePrizePool', deployer, [deployer,mockYieldSourceResult.address])
-  const ticketResult = await deployContract(deploy, 'Ticket', deployer, ["Ticket","TICK", TOKEN_DECIMALS, yieldSourcePrizePoolResult.address])
-  
+  const yieldSourcePrizePoolResult = await deployContract(deploy, 'YieldSourcePrizePool', deployer, [deployer, mockYieldSourceResult.address])
+  const ticketResult = await deployContract(deploy, 'Ticket', deployer, ["Ticket", "TICK", TOKEN_DECIMALS, yieldSourcePrizePoolResult.address])
+  const drawBufferResult = await deployContract(deploy, 'DrawBuffer', deployer, [deployer, DRAW_BUFFER_CARDINALITY])
+  const prizeDistributionBufferResult = await deployContract(deploy, 'PrizeDistributionBuffer', deployer, [deployer, PRIZE_DISTRIBUTION_BUFFER_CARDINALITY])
+  const drawCalculatorResult = await deployContract(deploy, 'DrawCalculator', deployer, [deployer, ticketResult.address, drawBufferResult.address, prizeDistributionBufferResult.address])
+  const prizeDistributorResult = await deployContract(deploy, 'PrizeDistributor', deployer, [deployer, ticketResult.address, drawCalculatorResult.address])
+  const prizeSplitStrategyResult = await deployContract(deploy, 'PrizeSplitStrategy', deployer, [deployer, yieldSourcePrizePoolResult.address])
+  const reserveResult = await deployContract(deploy, 'Reserve', deployer, [deployer, ticketResult.address])
+
+  // GET Contracts
   const yieldSourcePrizePool = await ethers.getContract('YieldSourcePrizePool')
+  const prizeSplitStrategy = await ethers.getContract('PrizeSplitStrategy')
 
   if (await yieldSourcePrizePool.getTicket() != ticketResult.address) {
     cyan('\nSetting ticket on prize pool...')
@@ -41,14 +48,6 @@ module.exports = async (hardhat) => {
     await tx.wait(1)
     green(`\nSet ticket!`)
   }
-
-  const drawBufferResult = await deployContract(deploy, 'DrawBuffer', deployer, [deployer, DRAW_BUFFER_CARDINALITY])
-  const prizeDistributionBufferResult = await deployContract(deploy, 'PrizeDistributionBuffer', deployer, [deployer, PRIZE_DISTRIBUTION_BUFFER_CARDINALITY])
-  const drawCalculatorResult = await deployContract(deploy, 'DrawCalculator', deployer, [deployer, ticketResult.address, drawBufferResult.address, prizeDistributionBufferResult.address])
-  const prizeDistributorResult = await deployContract(deploy, 'PrizeDistributor', deployer, [deployer,ticketResult.address, drawCalculatorResult.address])
-  const prizeSplitStrategyResult = await deployContract(deploy, 'PrizeSplitStrategy', deployer, [deployer, yieldSourcePrizePoolResult.address])
-  const reserveResult = await deployContract(deploy, 'Reserve', deployer, [deployer, ticketResult.address])
-  const prizeSplitStrategy = await ethers.getContract('PrizeSplitStrategy')
 
   if (await yieldSourcePrizePool.getPrizeStrategy() != prizeSplitStrategyResult.address) {
     cyan('\nSetting prize strategy on prize pool...')
@@ -66,9 +65,9 @@ module.exports = async (hardhat) => {
     green('Done!')
   }
 
-  const prizeFlushResult = await deployContract(deploy, 'PrizeFlush', deployer, [deployer,prizeDistributorResult.address,prizeSplitStrategyResult.address,reserveResult.address])
-  const drawCalculatorTimelockResult = await deployContract(deploy, 'DrawCalculatorTimelock', deployer, [deployer,drawCalculatorResult.address, DRAW_CALCULATOR_TIMELOCK])
-  await deployContract(deploy, 'L2TimelockTrigger', deployer, [deployer,drawBufferResult.address, prizeDistributionBufferResult.address,drawCalculatorTimelockResult.address])
+  const prizeFlushResult = await deployContract(deploy, 'PrizeFlush', deployer, [deployer, prizeDistributorResult.address, prizeSplitStrategyResult.address, reserveResult.address])
+  const drawCalculatorTimelockResult = await deployContract(deploy, 'DrawCalculatorTimelock', deployer, [deployer, drawCalculatorResult.address, DRAW_CALCULATOR_TIMELOCK])
+  await deployContract(deploy, 'L2TimelockTrigger', deployer, [deployer, drawBufferResult.address, prizeDistributionBufferResult.address, drawCalculatorTimelockResult.address])
   const l2TimelockTrigger = await ethers.getContract('L2TimelockTrigger')
   const reserve = await ethers.getContract('Reserve')
   const drawBuffer = await ethers.getContract('DrawBuffer')
@@ -88,7 +87,7 @@ module.exports = async (hardhat) => {
     await tx.wait(1)
     green(`Set reserve manager!`)
   }
-  
+
   /* ========================================= */
   // Phase 3 ---------------------------------
   // Set the manager(s) of the periphery smart contracts.
@@ -106,7 +105,7 @@ module.exports = async (hardhat) => {
     await tx.wait(1)
     green('Done!')
   }
-  
+
   if (await drawCalculatorTimelock.manager() != l2TimelockTrigger.address) {
     cyan(`\nSetting DrawCalculatorTimelock manager to ${l2TimelockTrigger.address}...`)
     const tx = await drawCalculatorTimelock.setManager(l2TimelockTrigger.address)
@@ -117,7 +116,7 @@ module.exports = async (hardhat) => {
   const prizeDistributionBuffer = await ethers.getContract('PrizeDistributionBuffer')
   if (await prizeDistributionBuffer.manager() != l2TimelockTrigger.address) {
     cyan(`\nSetting PrizeDistributionBuffer manager to ${l2TimelockTrigger.address}...`)
-    const tx =  await prizeDistributionBuffer.setManager(l2TimelockTrigger.address)
+    const tx = await prizeDistributionBuffer.setManager(l2TimelockTrigger.address)
     await tx.wait(1)
     green(`Done!`)
   }
