+++ package.json
@@ -2,7 +2,7 @@
   "name": "@pooltogether/v4-testnet",
   "version": "1.0.0-beta.4.1",
   "main": "index.js",
-  "repository": "git@github.com:pooltogether/v4-rinkeby.git",
+  "repository": "git@github.com:pooltogether/v4-testnet.git",
   "author": "Brendan Asselstine <mail.asselstine@gmail.com>",
   "license": "MIT",
   "scripts": {
@@ -58,7 +58,9 @@
     "node-emoji": "^1.11.0",
     "node-pty": "^0.10.1",
     "prompt": "^1.2.0",
-    "term.js": "^0.0.7"
+    "term.js": "^0.0.7",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.3"
   },
   "files": [
     "artifacts",
